PROGETTO UNIVERSITA'/requirements.txt
yfinance
pandas
numpy
matplotlib
seaborn
scipy
dwave-ocean-sdk 